requirements.txt
cffi==1.13.2
cryptography==2.8
ntlm-auth==1.4.0
pyasn1==0.4.8
pycparser==2.19
pysmb==1.1.28
six==1.14.0
smbprotocol==1.0.1
